describe hydration
