+++ packages/dds/tree/src/test/simple-tree/object.spec.ts
@@ -7,15 +7,11 @@ import { strict as assert } from "assert";
 
 import {
 	type ImplicitFieldSchema,
-	type NodeKind,
 	SchemaFactory,
 	type TreeFieldFromImplicitField,
-	type TreeNodeSchema,
 } from "../../simple-tree/index.js";
 
 import { hydrate, pretty } from "./utils.js";
-
-const schemaFactory = new SchemaFactory("Test");
 
 interface TestCase<TSchema extends ImplicitFieldSchema = ImplicitFieldSchema> {
 	name?: string;
@@ -380,168 +376,3 @@ const tcs: TestCase[] = [
 ];
 
 testObjectLike(tcs);
-
-const factory = new SchemaFactory("test");
-
-describe("Object-like-2", () => {
-	describe("setting an local field", () => {
-		it("throws TypeError in POJO emulation mode", () => {
-			const root = hydrate(schemaFactory.object("no fields", {}), {});
-			assert.throws(() => {
-				// The actual error "'TypeError: 'set' on proxy: trap returned falsish for property 'foo'"
-				// eslint-disable-next-line @typescript-eslint/no-explicit-any
-				(root as unknown as any).foo = 3;
-			}, "attempting to set an invalid field must throw.");
-		});
-
-		it("works in Customizable mode", () => {
-			class Custom extends schemaFactory.object("no fields", {}) {
-				public foo?: number;
-			}
-			const root = hydrate(Custom, {});
-			root.foo = 3;
-		});
-	});
-
-	describe("deep equality and types", () => {
-		it("types are ignored in POJO emulation mode", () => {
-			const a = hydrate(schemaFactory.object("a", {}), {});
-			const b = hydrate(schemaFactory.object("b", {}), {});
-			assert.deepEqual(a, {});
-			assert.deepEqual(a, b);
-		});
-
-		it("types are compared in Customizable mode", () => {
-			class A extends schemaFactory.object("a", {}) {}
-			class B extends schemaFactory.object("b", {}) {}
-			const a = hydrate(A, {});
-			const b = hydrate(B, {});
-			assert.notDeepEqual(a, {});
-			assert.notDeepEqual(a, b);
-			const a2 = hydrate(A, {});
-			assert.deepEqual(a, a2);
-		});
-	});
-
-	describe("supports setting", () => {
-		describe("primitives", () => {
-			function check<const TNode>(
-				schema: TreeNodeSchema<string, NodeKind, TNode>,
-				before: TNode,
-				after: TNode,
-			) {
-				describe(`required ${typeof before} `, () => {
-					it(`(${pretty(before)} -> ${pretty(after)})`, () => {
-						const Root = factory.object("", { value: schema });
-						const root = hydrate(Root, { value: before });
-						assert.equal(root.value, before);
-						root.value = after;
-						assert.equal(root.value, after);
-					});
-				});
-
-				describe(`optional ${typeof before}`, () => {
-					it(`(undefined -> ${pretty(before)} -> ${pretty(after)})`, () => {
-						const root = hydrate(
-							schemaFactory.object("", { value: schemaFactory.optional(schema) }),
-							{ value: undefined },
-						);
-						assert.equal(root.value, undefined);
-						root.value = before;
-						assert.equal(root.value, before);
-						root.value = after;
-						assert.equal(root.value, after);
-					});
-				});
-			}
-
-			check(schemaFactory.boolean, false, true);
-			check(schemaFactory.number, 0, 1);
-			check(schemaFactory.string, "", "!");
-		});
-
-		describe("required object", () => {
-			const Child = factory.object("child", {
-				objId: factory.number,
-			});
-			const Schema = factory.object("parent", {
-				child: Child,
-			});
-
-			const before = { objId: 0 };
-			const after = { objId: 1 };
-
-			it(`(${pretty(before)} -> ${pretty(after)})`, () => {
-				const root = hydrate(Schema, { child: before });
-				assert.equal(root.child.objId, 0);
-				root.child = new Child(after);
-				assert.equal(root.child.objId, 1);
-			});
-		});
-
-		describe("optional object", () => {
-			const Child = factory.object("child", {
-				objId: factory.number,
-			});
-			const Schema = factory.object("parent", {
-				child: factory.optional(Child),
-			});
-
-			const before = { objId: 0 };
-			const after = { objId: 1 };
-
-			it(`(undefined -> ${pretty(before)} -> ${pretty(after)})`, () => {
-				const root = hydrate(Schema, { child: undefined });
-				assert.equal(root.child, undefined);
-				root.child = new Child(before);
-				assert.equal(root.child.objId, 0);
-				root.child = new Child(after);
-				assert.equal(root.child.objId, 1);
-			});
-		});
-
-		describe.skip("required list", () => {
-			// const _ = new SchemaFactory("test");
-			// const list = _.fieldNode("List<string>", _.sequence(_.string));
-			// const parent = _.struct("parent", {
-			// 	list,
-			// });
-			// const schema = _.intoSchema(parent);
-			// const before: string[] = [];
-			// const after = ["A"];
-			// it(`(${pretty(before)} -> ${pretty(after)})`, () => {
-			// 	const root = getRoot(schema, { list: before });
-			// 	assert.deepEqual(root.list, before);
-			// 	root.list = after;
-			// 	assert.deepEqual(root.list, after);
-			// });
-		});
-
-		describe.skip("optional list", () => {
-			// const _ = new SchemaFactory("test");
-			// const list = _.fieldNode("List<string>", _.sequence(_.string));
-			// const parent = _.struct("parent", {
-			// 	list: _.optional(list),
-			// });
-			// const schema = _.intoSchema(parent);
-			// const before: string[] = [];
-			// const after = ["A"];
-			// it(`(undefined -> ${pretty(before)} -> ${pretty(after)})`, () => {
-			// 	const root = getRoot(schema, { list: undefined });
-			// 	assert.equal(root.list, undefined);
-			// 	root.list = before;
-			// 	assert.deepEqual(root.list, before);
-			// 	root.list = after;
-			// 	assert.deepEqual(root.list, after);
-			// });
-		});
-
-		describe("required map", () => {
-			// TODO
-		});
-
-		describe("optional map", () => {
-			// TODO
-		});
-	});
-});
